Create submission event listener
